+++ fixtures/license.ts
