+++ fixtures/src/index.js
