+++ fixtures/webpack.config.js
