+++ packages/test-v3/v3.test.ts
@@ -1,5 +1,5 @@
 import memoryfs from "memory-fs";
-import licenseComment from "../../__tests__/fixtures/license";
+import licenseComment from "../../fixtures/license";
 
 describe("webpack v3", () => {
   test("when license-info-webpack-plugin is used", done => {
